+++ lib/github.ml
@@ -36,6 +36,12 @@ module Parse = struct
     | _ when Bos_setup.String.is_prefix uri ~affix:"https://" ->
         user_from_regexp_opt uri "https://github\\.com/\\(.+\\)/.+\\(\\.git\\)?"
     | _ -> None
+
+  let ssh_uri_from_http uri =
+    (* if that fails, assume it's already an ssh uri *)
+    match String.cut ~sep:"https://github.com/" uri with
+    | Some ("", path) -> "git@github.com:" ^ path
+    | _ -> uri
 end
 
 (* Publish documentation *)
@@ -235,6 +241,51 @@ let assert_tag_exists ~dry_run tag =
   if Vcs.tag_exists ~dry_run repo tag then Ok ()
   else R.error_msgf "%s is not a valid tag" tag
 
+(* Ask the user then push the tag. Guess the ssh URI from the dev-repo.
+   This function can abort:
+   - The user answered "N" to pushing the tag
+   - The push failed
+
+   This function does nothing if the tag is already present on the remote and
+   point to the same ref. *)
+let push_tag ~dry_run ~yes ~dev_repo vcs tag =
+  let remote_has_tag_uptodate () =
+    if dry_run then Ok true (* Assume yes when [dry_run] is enabled *)
+    else
+      Vcs.commit_id ~dirty:false ~commit_ish:tag vcs >>= fun local_rev ->
+      Vcs.ls_remote ~dry_run vcs ~kind:`Tag ~filter:tag dev_repo >>= function
+      | [] -> Ok false
+      | (rev, _) :: _ when local_rev = rev -> Ok true
+      | (rev, _) :: _ ->
+          App_log.unhappy (fun l ->
+              l
+                "The tag %a is present on the remote but point to a different \
+                 commit (%a)."
+                Text.Pp.version tag Text.Pp.commit rev);
+          Ok false
+  in
+  remote_has_tag_uptodate () >>= function
+  | true -> Ok () (* No need to push, avoiding the need to guess the uri. *)
+  | false -> (
+      let uri = Parse.ssh_uri_from_http dev_repo in
+      Prompt.confirm_or_abort ~yes
+        ~question:(fun l ->
+          l "Push tag %a to %a?" Text.Pp.version tag Text.Pp.url uri)
+        ~default_answer:Yes
+      >>= fun () ->
+      App_log.status (fun l ->
+          l "Pushing tag %a to %a" Text.Pp.version tag Text.Pp.url uri);
+      match
+        Vcs.run_git_quiet vcs ~dry_run Cmd.(v "push" % "--force" % uri % tag)
+      with
+      | Ok () as ok -> ok
+      | Error (`Msg e) ->
+          R.error_msgf
+            "%s\n\
+             Pushing the tag failed, please push it manually and run the \
+             command again"
+            e )
+
 let publish_distrib ?token ?distrib_uri ~dry_run ~msg ~archive ~yes p =
   (match distrib_uri with Some uri -> Ok uri | None -> Pkg.infer_distrib_uri p)
   >>= fun uri ->
@@ -247,28 +298,19 @@ let publish_distrib ?token ?distrib_uri ~dry_run ~msg ~archive ~yes p =
   Vcs.get () >>= fun vcs ->
   Pkg.tag p >>= fun tag ->
   check_tag ~dry_run vcs tag >>= fun () ->
-  dev_repo p >>= fun upstr ->
-  Prompt.(
-    confirm_or_abort ~yes
-      ~question:(fun l ->
-        l "Push tag %a to %a?" Text.Pp.version tag Text.Pp.url upstr)
-      ~default_answer:Yes)
-  >>= fun () ->
-  App_log.status (fun l ->
-      l "Pushing tag %a to %a" Text.Pp.version tag Text.Pp.url upstr);
-  Vcs.run_git_quiet vcs ~dry_run Cmd.(v "push" % "--force" % upstr % tag)
-  >>= fun () ->
+  dev_repo p >>= fun dev_repo ->
+  push_tag ~dry_run ~yes ~dev_repo vcs tag >>= fun () ->
   (match token with Some t -> Ok t | None -> Config.token ~dry_run ())
   >>= fun token ->
   Prompt.(
     confirm_or_abort ~yes
       ~question:(fun l ->
-        l "Create release %a on %a?" Text.Pp.version tag Text.Pp.url upstr)
+        l "Create release %a on %a?" Text.Pp.version tag Text.Pp.url dev_repo)
       ~default_answer:Yes)
   >>= fun () ->
   App_log.status (fun l ->
       l "Creating release %a on %a via github's API" Text.Pp.version tag
-        Text.Pp.url upstr);
+        Text.Pp.url dev_repo);
   curl_create_release ~token ~dry_run tag msg user repo >>= fun id ->
   App_log.success (fun l -> l "Succesfully created release with id %d" id);
   Prompt.(