+++ tests/bin/no_doc/run.t
@@ -196,9 +196,6 @@ We do the whole process, calling publish doc implicitely should succeed
     [-] Publishing to github
     -: exec: git --git-dir .git rev-parse --verify 0.1.0
     -: exec: git --git-dir .git rev-parse --verify 0.1.0
-    [?] Push tag 0.1.0 to git@github.com:foo/whatever.git? [Y/n]
-    [-] Pushing tag 0.1.0 to git@github.com:foo/whatever.git
-    -: exec: git --git-dir .git push --force git@github.com:foo/whatever.git 0.1.0
     ...
     [?] Create release 0.1.0 on git@github.com:foo/whatever.git? [Y/n]
     [-] Creating release 0.1.0 on git@github.com:foo/whatever.git via github's API